scripts/requirements.txt
accelerate
datasets
elevenlabs
ffmpeg-python
Flask
google-generativeai
nltk
numpy<2.0
opencv-python
pandas
pyaudio
python-dotenv
requests
scikit-learn
torch
transformers
twelvelabs
yt-dlp
